requirements.txt
Flask>=2.3.0
Flask-RESTful==0.3.9
pytest==7.1.2
requests==2.26.0
Flask-SQLAlchemy>=3.0.0
openai==1.61.1
marshmallow==3.19.0
python-dotenv==1.0.1        